nyaybandhu-legal-chatbot/backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
python-dotenv
google-generativeai
langchain_community
langchain_huggingface
chromadb
